+++ lib/eva/lm/macro_builtins.py
@@ -1,0 +1,7 @@
+def alias(alias_obj, src_obj):
+    g = globals()
+    try:
+        g[alias_obj] = g[src_obj]
+        return True
+    except:
+        return False